Backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
supabase==2.0.2
postgresql-adapter==0.1.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Task queue and caching
celery==5.3.4
redis==5.0.1

# Web scraping
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
lxml==4.9.3

# Data processing and analysis
pandas==2.1.3
numpy==1.25.2

# Export functionality
openpyxl==3.1.2
xlsxwriter==3.1.9

# HTTP and CORS
httpx==0.25.2
fastapi-cors==0.0.6

# Validation and serialization
marshmallow==3.20.1

# Logging and monitoring
loguru==0.7.2

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Rate limiting
slowapi==0.1.9

# WebDriver management
webdriver-manager==4.0.1